pyproject.toml
[tool.poetry]
name = "laptime-sim"
version = "0.1.0"
description = ""
authors = ["JanJaapvandeVelde <janjaap.vandevelde@gmail.com>"]
readme = "README.md"
packages = [{include = "laptime_sim"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.0"
geopandas = "^0.14.0"
streamlit-folium = "^0.15.0"
plotly = "^5.17.0"
folium = "^0.14.0"
matplotlib = "^3.8.0"
mapclassify = "^2.6.1"
pyogrio = "^0.7.1"
shapelysmooth = "^0.1.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.pytest.ini_options]
addopts = ["--import-mode=importlib",]
pythonpath = [".", "tests", "laptime_sim"]
